requirements.txt
atlassian-python-api==3.36.0
certifi==2023.5.7
charset-normalizer==3.1.0
Deprecated==1.2.13
idna==3.4
mysql-connector-python==8.0.33
oauthlib==3.2.2
protobuf==3.20.3
python-dotenv==1.0.0
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
urllib3==2.0.2
wrapt==1.15.0
